+++ client-library/Protocol/GameState/ItemCard.cs
@@ -8,8 +8,8 @@ using Nethereum.ABI.FunctionEncoding.Attributes;
  * State associated with an card that is part of a players hand in battle
  */
 public record ItemCard {
-    [Parameter("uint8", "cardId", 1)]
-    public uint CardId { get; set; }
+    [Parameter("uint256", "cardId", 1)]
+    public BigInteger CardId { get; set; }
 
     [Parameter("address", "definition", 2)]
     public string Definition { get; set; }
